refactor: application interaction route table (v3.1.10 part 2/7)
Mirror the ticket interaction split for applications. The 467-LOC
`handleApplicationInteraction` if-ladder collapses to a 6-LOC dispatch
into a route table at `src/events/application/interactionRoutes.ts`
that keys on `customId` (exact and prefix) per interaction type.

Branch bodies extracted into:
- `apply.ts` — applyButton, ageVerifyYesButton, ageVerifyNoButton,
  cancelApplicationButton, submitApplicationModal, plus the private
  showApplicationModal helper that's shared between apply and ageVerifyYes
- `close.ts` — adds closeApplicationButton, confirmCloseApplication,
  cancelCloseApplication next to existing applicationCloseEvent

Behavior unchanged. 1071 tests still passing. Build + biome clean.
