+++ src/events/application/close.ts
@@ -1,5 +1,8 @@
 import {
+  ActionRowBuilder,
+  ButtonBuilder,
   type ButtonInteraction,
+  ButtonStyle,
   type Client,
   type ForumChannel,
   type ForumThreadChannel,
@@ -146,4 +149,39 @@ export const applicationCloseEvent = async (client: Client, interaction: ButtonI
       },
     );
   }
+};
+
+export const closeApplicationButton = async (_client: Client, interaction: ButtonInteraction) => {
+  enhancedLogger.debug(`Button: close_application`, LogCategory.COMMAND_EXECUTION, {
+    userId: interaction.user.id,
+    guildId: interaction.guildId,
+  });
+
+  const confirmRow = new ActionRowBuilder<ButtonBuilder>().addComponents(
+    new ButtonBuilder().setCustomId('confirm_close_application').setLabel(tl.closingL).setStyle(ButtonStyle.Danger),
+    new ButtonBuilder().setCustomId('cancel_close_application').setLabel(tl.cancelL).setStyle(ButtonStyle.Secondary),
+  );
+
+  await interaction.reply({
+    content: tl.confirm,
+    components: [confirmRow],
+    flags: [MessageFlags.Ephemeral],
+  });
+};
+
+export const confirmCloseApplication = async (client: Client, interaction: ButtonInteraction) => {
+  enhancedLogger.debug(`Button: confirm_close_application`, LogCategory.COMMAND_EXECUTION, {
+    userId: interaction.user.id,
+    guildId: interaction.guildId,
+  });
+  await interaction.update({ content: tl.closing, components: [] });
+  await applicationCloseEvent(client, interaction);
+};
+
+export const cancelCloseApplication = async (_client: Client, interaction: ButtonInteraction) => {
+  enhancedLogger.debug(`Button: cancel_close_application`, LogCategory.COMMAND_EXECUTION, {
+    userId: interaction.user.id,
+    guildId: interaction.guildId,
+  });
+  await interaction.update({ content: tl.cancel, components: [] });
 };
